requirements.txt
aiohttp==3.5.4
aiopg==0.16.0
aioredis==1.3.0
databases==0.2.5
locustio==0.11.0
passlib==1.7.1
bcrypt==3.1.7
asyncpg==0.18.3
PyJWT==1.7.1
python-multipart==0.0.5